add math dynamics base class
